Add a callback_url functionality to notification schema.
Adds a new field -callback_url- in notification schema and support
for it with TestJob field values substitutions.
Implements LAVA-927 (https://projects.linaro.org/browse/LAVA-927).

[Matt] Added support for pulling auth token from XMLRPC tokens for
submitting user with matching description
Added content-type option to respond in JSON.
Added auth-token to authorization header and fixed up the
docs.
Stop trying to send email if only POST callback is requested, but do
send email if a recipient is also requested.
Added a 'finished' status option that will send notifications on
canceled, complete or incomplete.
Added a status_string property to TestJob that is the string
representation of status. It's a more useful output than an integer.
Use {THING} as substitution to match LAVA elsewhere and not confuse
jinja with {{ }}

Change-Id: Ic567628eab90f1926ba59aed5bfe5a037f118c62