dispatcher-master: put the logging socket handling in a separate thread
The dispatcher-master was mainly a loop that did three things:
  * polling a "logging" queue for job logs
    -> one message was unqueued per iteration
  * polling a "controller" queue to sync with the slave(s)
    -> one message was unqueued per iteration
  * processing the scheduled jobs to send them to the slave(s)
    -> many messages were to be queued by the slave(s), depending on the number
    of jobs processed

A problem occurred when many jobs were scheduled at the same time: the
dispatcher-master sent all the jobs to the slave(s), via START signals, and the
slave(s) correctly answered with START_OK. But during the time taken to send all
the jobs, many START_OK were queued, and in the next loop iteration, the
dispatcher-master unqueued only one START_OK, before reprocessing all the
scheduled jobs (minus one, which has been marked as Running by the START_OK),
making the slave(s) to generate again many START_OK.

This ended up with the dispatcher-master having lot of START_OK in queue, making
it unable to process the coming END messages of the slave(s), thus marking the
jobs as finished (Complete or Incomplete). Also, it took a lot of time to mark
the jobs as Running, and sometimes they even didn't passed to that state before
being marked as finished, which is really inconvenient for the web UI.

A solution has been found by putting the logging socket handler in a separate
thread, and then rework the main thread not to make it generate messages while
its incoming queue still has messages to deal with.

This first patch puts the processing of the logging message queue in a thread, to
achieve better reactivity in the main thread that handles the commands.

Change-Id: I72091ca8f73c0259dbefd887b0a0c3bef64b377f
Signed-off-by: Florent Jacquet <florent.jacquet@free-electrons.com>